+++ script/deploy/oracles/DeployMockAggregators.s.sol
@@ -20,23 +20,12 @@ contract DeployMockAggregators is TrebScript {
         IMentoConfig config = Config.get();
         Senders.Sender storage deployer = sender("deployer");
 
-        IMentoConfig.MockAggregatorConfig[] memory aggConfigs = config
-            .getMockAggregatorConfigs();
+        IMentoConfig.MockAggregatorConfig[] memory aggConfigs = config.getMockAggregatorConfigs();
 
-        for (uint i = 0; i < aggConfigs.length; i++) {
-            address aggAddy = deployer
-                .create3("MockChainlinkAggregator")
-                .setLabel(aggConfigs[i].label)
-                .deploy(
-                    abi.encode(
-                        aggConfigs[i].description,
-                        aggConfigs[i].decimals,
-                        deployer.account
-                    )
-                );
-            MockChainlinkAggregator agg = MockChainlinkAggregator(
-                deployer.harness(aggAddy)
-            );
+        for (uint256 i = 0; i < aggConfigs.length; i++) {
+            address aggAddy = deployer.create3("MockChainlinkAggregator").setLabel(aggConfigs[i].label)
+                .deploy(abi.encode(aggConfigs[i].description, aggConfigs[i].decimals, deployer.account));
+            MockChainlinkAggregator agg = MockChainlinkAggregator(deployer.harness(aggAddy));
             agg.setExternalProvider(config.mockAggregatorReporter());
             agg.report(aggConfigs[i].initialReport, block.timestamp);
         }
